fix: address F03 review findings

+++ xiaou-starter/src/main/java/com/xiaou/web/organization/OrganizationMemberListResponse.java
@@ -4,5 +4,5 @@ import java.util.List;
 
 public record OrganizationMemberListResponse(
         List<OrganizationMemberResponse> items,
-        long total) {
+        int total) {
 }
